feat: add common serialize package
